+++ MedicalSystemApi/Controllers/BillController.cs
@@ -1,6 +1,7 @@
 ﻿using MedicalSystemApi.Models.DTOs.Bill;
 using MedicalSystemApi.Models.Entities;
 using MedicalSystemApi.Services.Interfaces;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace MedicalSystemApi.Controllers
@@ -9,18 +10,20 @@ namespace MedicalSystemApi.Controllers
     [ApiController]
     [ApiVersion("2.0")]
 
-    public class BillController : ControllerBase
+    public class BillController(IBillService billService) : ControllerBase
     {
-        private readonly IBillService _billService;
-
-        public BillController(IBillService billService)
-        {
-            _billService = billService;
-        }
+        private readonly IBillService _billService = billService;
 
         [HttpGet("AllBills")]
+        [Authorize(Roles = "SuperAdmin, Admin")]
+        [ResponseCache(Duration = 60, Location = ResponseCacheLocation.Client, NoStore = false)]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             try
             {
                 var billList = await _billService.GetAllAsync();
@@ -33,8 +36,15 @@ namespace MedicalSystemApi.Controllers
         }
 
         [HttpGet("{id:int}")]
+        [Authorize(Roles = "SuperAdmin, Admin")]
+        [ResponseCache(Duration = 60, Location = ResponseCacheLocation.Client, NoStore = false)]
         public async Task<IActionResult> GetById(int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             try
             {
                 if (!ModelState.IsValid)
@@ -52,6 +62,7 @@ namespace MedicalSystemApi.Controllers
         }
 
         [HttpPost("AddBill")]
+        [Authorize(Roles = "SuperAdmin")]
         public async Task<IActionResult> Add(CreateBillDto createBillDto)
         {
             try
@@ -71,6 +82,7 @@ namespace MedicalSystemApi.Controllers
         }
 
         [HttpPut("{id}")]
+        [Authorize(Roles = "SuperAdmin")]
         public async Task<IActionResult> Update(int id, [FromBody] UpdateBillDto updateBillDto)
         {
             try
@@ -90,8 +102,14 @@ namespace MedicalSystemApi.Controllers
         }
 
         [HttpDelete("{id}")]
+        [Authorize(Roles = "SuperAdmin")]
         public async Task<IActionResult> Delete(int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             try
             {
                 await _billService.DeleteAsync(id);
@@ -104,6 +122,8 @@ namespace MedicalSystemApi.Controllers
         }
 
         [HttpGet("patient/{patientId}")]
+        [Authorize(Roles = "SuperAdmin, Admin")]
+        [ResponseCache(Duration = 60, Location = ResponseCacheLocation.Client, NoStore = false)]
         public async Task<ActionResult<Bill>> GetBillsByPatient(int patientId)
         {
             try
@@ -123,12 +143,39 @@ namespace MedicalSystemApi.Controllers
         }
 
         [HttpPut("update-total/{billId}")]
+        [Authorize(Roles = "SuperAdmin")]
         public async Task<IActionResult> UpdateTotalAmount(int billId, [FromBody] decimal newTotalAmount)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             try
             {
                 await _billService.UpdateTotalAmountAsync(billId, newTotalAmount);
                 return NoContent();
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+        [HttpGet("Filtering")]
+        [Authorize(Roles = "SuperAdmin, Admin")]
+        [ResponseCache(Duration = 60, Location = ResponseCacheLocation.Client, NoStore = false)]
+        public async Task<IActionResult> GetFilteredBills([FromQuery] BillFilterDto filterDto)
+        {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            try
+            {
+                var bills = await _billService.GetFilteredBillsAsync(filterDto);
+                return Ok(bills);
             }
             catch (Exception ex)
             {
